% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_climwin.R
\name{analyse_climwin}
\alias{analyse_climwin}
\title{Analyze the results of climate window analysis}
\usage{
analyse_climwin(ID, biol_data, out_dir = "output_climwin",
  randwin = FALSE, metric = "C")
}
\arguments{
\item{ID}{Numeric giving a unique ID of the current dataset for
a given population and species.}

\item{biol_data}{Data frame with trait data for a given
population and species.}

\item{out_dir}{Character specifying the library on the path where
the results of climate window analysis were stored.}

\item{randwin}{Logical (TRUE/FALSE) specifying whether \code{\link[climwin]{randwin}}
was run together with \code{\link[climwin]{slidingwin}}.}

\item{metric}{Character specifying 'AIC' or 'C'. This define whether
a value of PDAIC or Pc will be returned. For more information
see the same option in the function \code{\link[climwin]{pvalue}}}
}
\value{
If the results are not likely to be an issue of overfitting,
then the function returns a tibble with four columns: "ID"
- study id, "Species" - study species, "pvalue" - the pvalue from
the test assessing whether the climate signal is obtained by chance only,
and "data_res" - a data frame with all the information required
for running the SEM. If the results are likely an issue of
overfitting, then the returned tibble contains only first three of
the above-mentioned columns.
}
\description{
\code{analyse_climwin} analyses the results obtained
with climate window analysis (function \code{\link{climwin_proc}})
for the specified study ID
}
\examples{
dat_birds <- read.csv('./data-raw/Test_european_birds.csv')
t_anal <- analyse_climwin(ID = 1, biol_data = dat_birds,
                          out_dir = 'output_climwin',
                          randwin = TRUE, metric = 'C')
}
