% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPaths_descr.R
\docType{data}
\name{dataPaths}
\alias{dataPaths}
\title{dataPaths}
\format{
An object of class \code{data.frame} with 840 rows and 39 columns.
}
\usage{
dataPaths
}
\description{
subset of the data on which SEM analyses were run,
and path coefficients estimated
#' @format A data frame with 840 rows and 39 variables:
\describe{
  \item{\code{Species}}{factor Latin species name}
  \item{\code{Response}}{factor Name of the variable used
  as a response in the SEM}
  \item{\code{Predictor}}{factor Name of the variable used
  as a predictor in the SEM}
  \item{\code{Estimate}}{double Estimated coefficient value}
  \item{\code{Std.Error}}{double Standard error of the estimate}
  \item{\code{DF}}{double Degrees of freedom}
  \item{\code{Crit.Value}}{double COLUMN_DESCRIPTION}
  \item{\code{P.Value}}{double COLUMN_DESCRIPTION}
  \item{\code{Std.Estimate}}{double Standardised estimate,
  here is the same as the estimate itself since analyses are
  run on standardised data}
  \item{\code{ID}}{integer Study ID}
  \item{\code{Study_Authors}}{factor Authors of the study}
  \item{\code{Location}}{factor Location of the study}
  \item{\code{Country}}{factor Country}
  \item{\code{Continent}}{factor Continent}
  \item{\code{Longitude}}{double Longitude}
  \item{\code{Latitude}}{double Latitude}
  \item{\code{Taxon}}{factor Taxon (can be bird, reptile,
  fish or mammal)}
  \item{\code{BirdType}}{factor Specifies whether the
  bird is a Seabird or not}
  \item{\code{Trait_Categ}}{factor Trait category, can
  be either Morphological or Phenological}
  \item{\code{Trait}}{factor Specific type of the trait
  measured in the study, as reported in the publication}
  \item{\code{Demog_rate_Categ}}{factor Demographic rate category,
  one of: Reproduction, Recruitment, Survival}
  \item{\code{Demog_rate}}{factor Demographic rate measured}
  \item{\code{Count}}{factor Specifies whether the
  population size was obtained by counts ('Y') or not ('N')}
  \item{\code{Nyears}}{integer Duration of the study, extracted
  with [prep_subset()] function}
  \item{\code{WinDur}}{double Duration of the climatic window}
  \item{\code{Ref.day}}{double Reference day for sliding
  window analysis}
  \item{\code{Ref.month}}{double Reference month for sliding
  window analysis}
  \item{\code{WindowClose}}{double The closest number of time
  intervals (set by cinterval) back from the reference day, used
  in sliding window analysis. See [climwin_proc()] function
  for more details}
  \item{\code{deltaAIC}}{double Delta AIC for the used model
  in sliding window analysis}
  \item{\code{Pvalue}}{double P deltaAIC value from sliding
  window analysis}
  \item{\code{weights}}{logical Specifies whether weights were used
  when fitting the SEM}
  \item{\code{DD}}{integer Specifies whether population size was
  included as a covariate in the SEM}
  \item{\code{corr}}{logical Specifies whether autocorrelation
  was included in SEM}
  \item{\code{AIC.AIC}}{AIC output for the fitted SEM}
  \item{\code{AIC.AICc}}{AICc output for the fitted SEM}
  \item{\code{Trait_ageClass}}{factor Specifies for which age
    class trait measurements are taken}
  \item{\code{Trait_ageClass}}{factor Specifies
  the age class for which the trait was measured}
  \item{\code{WeathQ}}{factor Specifies
  whether climate was extracted from the location or
  from a nearby fastland (see Methods for details)}
  \item{\code{Relation}}{factor Specifies for which relation
  the path coefficient was estimated with the SEM}
  \item{\code{GenLength_y_IUCN}}{double Generation time in years}
  \item{\code{TraitTyoe}}{factor Specifies the detailed category for
  the phenotypic trait}
}
}
\keyword{datasets}
