% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extr_coefs}
\alias{extr_coefs}
\title{Extract path coefficients from the output of \code{\link{fit_all_meta}} function}
\usage{
extr_coefs(obj, Type_EfS)
}
\arguments{
\item{obj}{An object returned by the  \code{\link{fit_all_meta}} function}

\item{Type_EfS}{Character specifying for which relation to extract the path coefficients,
These relations reflect different pathways in the fitted SEM, for example 'Demog_rate_mean<-det_Clim',
'Demog_rate_mean<-Pop_mean', 'Demog_rate_mean<-Trait_mean'. For more details see \code{\link{fit_meta}}}
}
\value{
A dataframe that contains 26 columns including the required path coefficients and
the associated meta-data.
}
\description{
\code{extr_coefs} extracts the path coefficients for a specified relation.
}
\examples{
 ## come up with a simpler example???
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
meta_Phen_Surv_neg <- fit_all_meta(data_MA = Coefs_Aut_neg,
Clim = 'Temperature',
Demog_rate = 'Survival',
Trait_categ = 'Phenological',
Covar = NULL,
COV = 'Pvalue + WeathQ',
sel = 'Temp_Neg_Phen_Surv_Cov',
folder_name = './output_overall_sig/',
colr = c('black'),
optimize = rep('uobyqa', 11))
test <- extr_coefs(data = meta_Phen_Surv_neg, Relation = 'Tot_DemRate<-det_Clim')
}
