% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_accountphylo.R
\name{fit_all_acountphylo}
\alias{fit_all_acountphylo}
\title{Fit all meta-analytical models for a specified subselection of
trait and climate using selected phylogeny}
\usage{
fit_all_acountphylo(
  data_MA,
  vtree_folder,
  ind,
  Trait_categ = "Phenological",
  Clim = "Temperature"
)
}
\arguments{
\item{data_MA}{Data frame containing, for each study, the effect size estimates
for each pathway from the SEM analyses. This data frame also contains meda-data
(e.g. sutdy species, study location, continent, life history traits of the species),
that are needed to fit the mixed-effect model.}

\item{vtree_folder}{Character specifying the folder where the phylogenetic
tree is to be read from. The tree is expected to be in Newick format and be
named as "vertX.tre", where X is an integer (from 1 to 100).}

\item{ind}{An integer specifying what phylogenetic tree to be read from the
folder given by vtree_folder.}

\item{Trait_categ}{Character specifying the level of the trait on which to subset the data.}

\item{Clim}{Character specifying the level of the climatic variable on which to subset
the data.}
}
\value{
Returns a tibble that includes the estimated across-study effect sizes,
their standard errors, their significance and the AIC for each fitted mixed-effects
model (fitted with or without phylogeny accounted for by using
phylogenetic variance-covariance matrix). This tibble also
includes variances for random effects.
}
\description{
\code{fit_all_acountphylo} fits all mixed-effects meta-analytical models
(per each relation in the SEM), to extract global effect sizes across
the studies in a specified subset of data as defined by trait, climate
and using a given phylogenetic tree
}
\examples{
phenT <- fit_all_acountphylo(data_MA = dataPaths,
                             vtree = "./data/phylogenies",
                             ind = 2, Trait_categ = "Phenological",
                             Clim = "Temperature")

}
