% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_meta_phylo.R
\name{fit_meta_phylo}
\alias{fit_meta_phylo}
\title{Fit a meta-analytical model to assess the global effects across the studies}
\usage{
fit_meta_phylo(
  data_MA,
  Type_EfS = "Trait_mean<-det_Clim",
  Cov_fact = NULL,
  COV = NULL,
  optimize = "uobyqa",
  DD = "n_effectDGR",
  simpleSEM = FALSE,
  Trait = FALSE,
  A = Mat_phylo,
  des.matrix = "treatm.contrasts"
)
}
\arguments{
\item{data_MA}{Data frame containing, for each study, the effect size estimates
for each pathway from the SEM analyses. This data frame also contains meda-data
(e.g. sutdy species, study location, continent, life history traits of the species),
that are needed to fit the mixed-effect model.}

\item{Type_EfS}{Character specifying which effect size type to use for a meta-analysis.
These types of effect sizes reflect different pathways in the fitted SEM. Possible are:
'Demog_rate_mean<-det_Clim', 'Demog_rate_mean<-Pop_mean', 'Demog_rate_mean<-Trait_mean',
'GR<-Demog_rate_mean',  'GR<-det_Clim', 'GR<-Pop_mean', 'Ind_DemRate<-det_Clim',
'Ind_GR<-det_Clim', 'Tot_DemRate<-det_Clim', 'Tot_GR<-det_Clim', and 'Trait_mean<-det_Clim'.}

\item{Cov_fact}{A character specifying the name of the categorical variable to be included
as a fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case no
categorical variables are included and the overall global effect size is estimated.}

\item{COV}{A character specifying the name (or names, separated by '+') of the continuous
variable to be included as a fixed-effect covariate in the meta-analysis. Defaults to NULl,
in which case no continuous covariates are included and the overall global effect size
is estimated.}

\item{optimize}{A character specifying the name of the optimizer to use. For more details
see argument 'control = ' for the available optimizers in \code{\link[metafor]{rma.mv}}.
Defaults to 'nlminb' because this optimizer was the most successful for model convergence
for our datasets.}

\item{DD}{Character specifying how to account for density dependence by
including population size as additional explanatory in the SEM. Possibilities
are: 'none' - no inclusion of density dependence, 'n_effectGR' - effect
of population size on growth rate, 'n_effectD' - effect of population size
on demographic rate, 'n_effectDGR' - effect of population size on demographic
rate and growth rate. Only options 'n_effectGR' and 'none' are possible for
simple structure of SEM, in which climate effect on GR i sonly mediated by trait
(i.e. \code{simpleSEM = TRUE}).}

\item{simpleSEM}{Logical (TRUE/FALSE) specifying that the simple SEM
structure will be fitted in which climate effect on GR is only mediated
by trait, demographic rates are not considered. Defaults to FALSE.}

\item{Trait}{Logical (TRUE/FALSE) specifying whether to include
the direct impact of trait on GR in the SEM fitted using a complex
structure, i.e. climate effect on GR is mediated by both trait and
demographic rate. Is applicable only if \code{simpleSEM = FALSE}.
Defaults to FALSE.}

\item{A}{A variance-covariance matrix based on phylogeny.}

\item{des.matrix}{A character specifying what type of design matrix to use if the factor
is included as one of the explanatory variables. By default treatment contrasts is used
 (des.matrix = "treatm.contrasts") but the user can also request identity matrix by
 specifying "identity".}
}
\value{
Returns a tibble that includes five data frames:
1. 'data' - the data frame with the estimate
of the global effect size, its standard error, its significance, the AIC of the fitted
mixed-effects model and the variances explained by random effects;
2. 'data_Efs' - the data frame with the effect sizes and their standard errors for each study;
3. 'heter_mod' - the data frame with estimates of heterogeneity
typically extracted for meta-analysis;
4. ML_mod - results of the meta-analysis correcting for phylogeny (if that
improves mode lit to the data) and fitted with ML;
5. REML_mod - results of the meta-analysis correcting for phylogeny (if that
improves mode lit to the data) and fitted with REML.
The AIC and the p values are obtained from the models fitted with ML, whereas
the parameter estimates from the models fitted using REML.
}
\description{
\code{fit_meta_phylo} fits mixed-effects meta-analytical model to extract a global
effect size across the studies
}
\examples{
# prepare dataset, select only studies with phenological traits
Coefs_phenClim <- subset(dataPaths, Relation == 'Trait_mean<-det_Clim' &
Trait_Categ == 'Phenological')
Coefs_phenClim <- Coefs_phenClim \%>\%
                  dplyr::mutate(Species = dplyr::case_when(
                         Species == 'Cyanistes caeruleus' ~ 'Parus caeruleus',
                         Species == 'Thalasseus sandvicensis' ~ 'Sterna sandvicensis',
                         Species == 'Setophaga caerulescens' ~ 'Dendroica caerulescens',
                         Species == 'Thalassarche melanophris' ~ 'Thalassarche melanophrys',
                         Species == 'Ichthyaetus audouinii' ~ 'Larus audouinii',
                         Species == 'Stercorarius maccormicki' ~ 'Catharacta maccormicki',
                         TRUE ~ Species))

Coefs_phenClim$Species <- unlist(lapply(1:nrow(Coefs_phenClim), FUN = function(x){
  binary <- strsplit(as.character(Coefs_phenClim$Species[x]), " ")
  Underscore <- paste(binary[[1]][1], binary[[1]][2], sep = "_")}))
Coefs_phenClim$Sp_phylo <- Coefs_phenClim$Species

test_noCovar <- fit_meta_phylo(data_MA = Coefs_phenClim,
                              Type_EfS = 'Trait_mean<-det_Clim',
                              Cov_fact = NULL, COV = NULL,
                              DD = 'n_effectGR',
                              simpleSEM = TRUE,
                              A = phyloMat)
# meta-analysis with a quantitative covariate
test_PValCovar <- fit_meta_phylo(data_MA = Coefs_phenClim,
                              Type_EfS = 'Trait_mean<-det_Clim',
                              Cov_fact = NULL, COV = 'Pvalue',
                              DD = 'n_effectGR',
                              simpleSEM = TRUE,
                              A = phyloMat)

test_WeathQ <- fit_meta_phylo(data_MA = Coefs_phenClim,
                                   Type_EfS = 'Trait_mean<-det_Clim',
                                   Cov_fact = 'WeathQ',
                                   COV = NULL,
                                   DD = 'n_effectGR',
                                   simpleSEM = TRUE,
                                   A = phyloMat)
test_WeathQ
}
