% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mod_beta.R
\name{fit_mod_beta}
\alias{fit_mod_beta}
\title{Fit beta regression model to the proportional contribution data}
\usage{
fit_mod_beta(prop_data, table_name, explanatory = NULL)
}
\arguments{
\item{prop_data}{Data frame containing for each study ID proportional
contribution of the trait-mediated effect either to the demographic rate
or to the growth rate (previously prepared by the user)}

\item{table_name}{A string specifying the path and the name of the table to be created.}

\item{explanatory}{Character specifying the name of the additional explanatory
(i.e. species-specific trait) to be included in the model.}
}
\value{
A list of length 2, with the first object containing
the fit statistics for each explanatory in the model (performed
with Wald test), and the secon dobject containing the fitted full model.
}
\description{
\code{fit_mod_beta} fits beta regression model either with or
without an additional explanatory (species trait), as requested
}
\examples{
prop_T_simple <- readRDS(file = './output_all_simpleSEM/all_prop_EffectTempOnTrait_GroupsTraitDem.RDS')
prop_T_simple <- prop_T_simple \%>\%
      dplyr::mutate(., Climatic_var = 'Temperature')
prop_P_simple <- readRDS(file = './output_all_simpleSEM/all_prop_EffectPrecipOnTrait_GroupsTraitDem.RDS')
prop_P_simple <- prop_P_simple \%>\%
     dplyr::mutate(., Climatic_var = 'Precipitation')
prop_sim <- rbind(prop_T_simple, prop_P_simple)
prop_long <- prop_sim \%>\%
     tidyr::gather(key = 'Path', value = 'Proportion',
     dplyr::contains('prop_'))
prop_IndGR <- prop_long \%>\%
     dplyr::filter(., Path == 'prop_ind_GR')
propGR_bet <- fit_mod_beta(prop_data = prop_IndGR,
                           table_name = './tables/PropGR_general',
                           explanatory = NULL)
propGR_bet[[1]]
mod_GR <- propGR_bet[[2]]
mod_GR

}
