% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mod_nonL.R
\name{fit_mod_nonL}
\alias{fit_mod_nonL}
\title{Fit mixed-effects models to predict GR by climate and trait}
\usage{
fit_mod_nonL(biol_data, ID, correlation = FALSE)
}
\arguments{
\item{biol_data}{Data frame with climate, trait and demographic rate
data for a given population and species.}

\item{ID}{Numeric giving a unique ID of the current dataset for
a given population and species.}

\item{correlation}{Logical (TRUE/FALSE) specifying whether to include
temporal autocorrelation with AR(1) structure. Defaults to FALSE.}
}
\value{
If all models fit well, returns a tibble containing a
data frame, the study ID, the trait category, and all four
fitted models (full quadratic, where a quadratic effect is
included for both climate and trait; quadratic climate, where
a quadratic effect is only included for climate; quadratic traits,
where a quadratic effect is only included for trait; and full
linear model where both climate and trait have a linear effect
on GR). The data frame contains the study ID, the trait category,
and AIC and df for each of the four models.
If at least one model fails to fit, returns a warning.

## Still add examples
}
\description{
\code{fit_mod_nonL} fits mixed-effects model to predict GR by
climate and trait, using varying structures of non-linear effects
for each of the predictors
}
