% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nonLinGR.R
\name{fit_nonLinGR}
\alias{fit_nonLinGR}
\title{Compare fit of mixed-effects models used to predict GR by non-linear effects of climate and trait}
\usage{
fit_nonLinGR(
  biol_data,
  ID,
  out_nonLinM,
  correlation = FALSE,
  standardize = FALSE
)
}
\arguments{
\item{biol_data}{Data frame with climate, trait and demographic rate
data for a given population and species.}

\item{ID}{Numeric giving a unique ID of the current dataset for
a given population and species.}

\item{correlation}{Logical (TRUE/FALSE) specifying whether to include
temporal autocorrelation with AR(1) structure. Defaults to FALSE.}
}
\value{
If all models fit well, returns a tibble, that, in addition
to the fields contained in the object returned by \code{fit_mod_nonL},
also contains the column (ModSt_minAIC) specifying the model with the
minimum AIC, and the column (minAICmod). The data frame, in
addition to the columns returned by \code{fit_mod_nonL}, also contains
columns ModSt_minAIC (a character specifying the 'best' model, i.e. the
one with the lowest AIC), BestMod2 (the second-best model as judged by AIC),
DeltAIC_best2 (a numeric giving delta AIC between the best and the
second-best model), DeltAIC_lin (a numeric specifying delta AIC of the
best model with the full linear model), and the coefficient estimates
for linear and quadratic (if applicable) effects of climate on trait,
for the best model.
If at least one model fails to fit, returns a warning.


## Still add examples
}
\description{
\code{fit_nonLinGR} prepares data in a way analogous to that used to fit SEM, fits
four mixed-effects models differing in how they incorporate quadratic effect of
climate and / or trait on GR, and compares these models (using AIC)
}
