% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SEM.R
\name{fit_SEM}
\alias{fit_SEM}
\title{Fit structural equation model and return the results}
\usage{
fit_SEM(
  biol_data,
  ID,
  out_SEM,
  DD = "none",
  weight = FALSE,
  correlation = FALSE,
  standardize = FALSE,
  detrend = FALSE
)
}
\arguments{
\item{biol_data}{Data frame with climate, trait and demographic rate
data for a given population and species.}

\item{ID}{Numeric giving a unique ID of the current dataset for
a given population and species.}

\item{out_SEM}{Character specifying the library on the path where
the results of structural equation analysis are stored.}

\item{DD}{Logical (TRUE/FALSE) specifying whether to account for density
dependence by including population size as additional explanatory of
population growth rate.}

\item{correlation}{Logical (TRUE/FALSE) specifying whether to include
temporal autocorrelation with AR(1) structure.}

\item{standardize}{Logical (TRUE/FALSE) specifying whether to
standardize (using z scores) the variables prior to fitting the model.}

\item{weights}{Logical (TRUE/FALSE) specifying whether to use inverse of
variances for traits and demographic rates as weights in the respective
models.}
}
\value{
A tibble with 15 columns that represent the metadata for a
given study ID and the results of SEM: Cstat, dTable, R2 of all
equations and path coefficients.
}
\description{
\code{fit_SEM} fits structural equation model with
requested model structure for the specified study ID and returns
the results to be used in meta-analysis
}
\examples{
biol_dat <- read.csv('./data-raw/Closed_Data_26_02.csv')
biol_stand <- convert_JulianDay(biol_data = biol_dat)
# example for ID = 1
subs <- droplevels(biol_stand[biol_stand$ID == 25, ])
test_fSEM <- readRDS(paste0('./output_forSEM_temp/', subs$ID[1], '_',
                    subs$Species[1], '_', subs$Location[1],
                    '_', subs$Trait[1], '_ForSEM',  '.RDS'))
dat <- test_fSEM$data_res[[1]]
# fit the model
test <- fit_SEM(biol_data = dat, ID = 1,
                out_SEM = 'output_SEM',
                DD = FALSE, weights = FALSE,
                correlation = FALSE,
                standardize = FALSE)  ## does not work for corr = TRUE yet

}
