% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_modPredGR.R
\name{fun_modPredGR}
\alias{fun_modPredGR}
\title{Predict GR by climate and trait from the fitted linear or mixed-effects model}
\usage{
fun_modPredGR(
  mods,
  ID,
  Trait_categ = "Phenological",
  det_Clim = 0,
  Year = 2000,
  Trait_mean = seq(-4, 4, by = 0.25),
  Pop_mean = 0,
  ME = FALSE
)
}
\arguments{
\item{mods}{A fitted mixed-effects or linear model with GR as a response and climate
and trait (as well as their quadratic effects, if desirable) as fixed effects. The linear
models are returned by the function \code{fit_nonLinGR}.}

\item{Trait_categ}{Character specifying the level of the trait, for which analyses were
conducted.}

\item{det_Clim}{The value(s) of det_Clim to be used for predicting GR.}

\item{Year}{The value(s) of year to be used for predicting GR, is needed only for
the mixed-effects model.}

\item{Trait_mean}{The value(s) of Trait_mean to be used for predicting GR.}

\item{Pop_mean}{The value(s) of Pop_mean to be used for predicting GR.}

\item{ME}{A Boolean specifying whether the supplied \code{mods} is a mixed-effects
model (TRUE) or a linear model (FALSE).}
}
\value{
A data frame with the requested values for all predictor variables and
the predicted GR values.

## Still add examples
}
\description{
\code{fun_modPredGR} predicts GR to the new data, whose columns are specified by the
user. The GR is predicted using the fitted linear or mixed-effects
model with trait and climate as predictors.
}
