% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lab_name.R
\name{generate_pdf_name}
\alias{generate_pdf_name}
\title{Generate a pdf name for the plot}
\usage{
generate_pdf_name(
  Relation = "GR<-det_Clim",
  Clim = "Temperature",
  Demog_rate = "Survival",
  Trait_categ = "Phenology"
)
}
\arguments{
\item{Relation}{Character specifying for which type of the effect size a meta-analysis
was conducted. Possible are: 'Demog_rate_mean<-det_Clim', 'Demog_rate_mean<-Pop_mean',
'Demog_rate_mean<-Trait_mean', 'GR<-Demog_rate_mean', 'GR<-det_Clim', 'GR<-Pop_mean',
'Ind_DemRate<-det_Clim', 'Ind_GR<-det_Clim', 'Tot_DemRate<-det_Clim',
'Tot_GR<-det_Clim', and 'Trait_mean<-det_Clim'. For more details see
\code{\link{fit_meta_phylo}}.}

\item{Clim}{Character specifying the level of the climatic variable, for which analyses
were conducted.}

\item{Demog_rate}{Character specifying the level of the demographic rate for which
analyses were conducted.}

\item{Trait_categ}{Character specifying the level of the trait, for which analyses were
conducted.}
}
\value{
A vector of length one with the pdf name.
}
\description{
\code{generate_pdf_name} prepares the pdf name for the forest
and 'relation' plots. This is an internal function
used by \code{\link{plot_forest}}.
}
