% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heterog.R
\name{get_heterog}
\alias{get_heterog}
\title{Assess model heterogeneity}
\usage{
get_heterog(mod, data)
}
\arguments{
\item{mod}{A fitted mixed-effects meta-analytical model.}

\item{data}{Data frame used to fit the meta-analytical model,
which contains effect sizes and the respective sampling variances.}
}
\value{
A tibble of the calculated heterogeneity metrics: Q reflects
the total amount of heterogeneity, Qp is the p value showing whether the observed
heterogeneity is larger than would be expected based on sampling variability,
I2 reflects the proportion of the total heterogeneity due to between-study
variance, ranges from 0 to 1, and I2_perRand is the proportion of the total
heterogeneity per random factor included in the model.
}
\description{
\code{get_heterog} computes the heterogeneity metrics for meta-analysis.
}
\details{
This function computes the heterogeneity metrics for the meta-analysis,
including I2, Q and p value associated with !
}
