% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{impute_median}
\alias{impute_median}
\title{Impute a median of all the values in the column}
\usage{
impute_median(data, column)
}
\arguments{
\item{data}{A dataframe containing the column for which the values have to be imputed.}

\item{Column}{A character specifying the name of the column for which the missing values
have to be imputed.}
}
\value{
A dataframe that contains imputed values for the column that had missing values
initially, or the original dataframe if the indicated column did not contain missing
values.
}
\description{
\code{imput_median} imputes the median of all the values in the indicated column.
}
\examples{
dat_birds <- read.csv('./data-raw/Test_european_birds.csv')
sub_1 <- droplevels(subset(dat_birds, ID == 1))
test <- impute_median(data = sub_1, column = 'Trait_SE')
sub_21 <- droplevels(subset(dat_birds, ID == 21))
test21 <- impute_median(data = sub_21, column = 'Trait_SE')
}
