% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_paths.R
\name{ind_path}
\alias{ind_path}
\title{Provide estimates of median and SE of an indirect (focal) pathway by bootstrap}
\usage{
ind_path(
  x,
  y,
  z = NULL,
  x.se,
  y.se,
  z.se = NULL,
  numrep = 10000,
  omega = NULL,
  omega.se = NULL
)
}
\arguments{
\item{x}{Numeric reflecting the first path coefficient estimate
in the indirect path}

\item{y}{Numeric reflecting the second path coefficient estimate
in the indirect path}

\item{z}{Numeric reflecting the third path coefficient estimate
in the indirect path, defaults to NULL, in which case
the indirect path consists of two pathways.}

\item{x.se}{Numeric reflecting the standard error for
the first path coefficient estimate.}

\item{y.se}{Numeric reflecting the standard error for
the second path coefficient estimate.}

\item{z.se}{Numeric reflecting the standard error for
the third path coefficient estimate, defaults to NULL for
the case of the indirect path consisting of two pathways.}

\item{numrep}{Numeric for the number of replicates to use in a
bootstrap.}

\item{omega}{Numeric reflecting the fourth path coefficient estimate
(applicable only if Trait = TRUE in \code{fit_SEM}). Defaults to NULL.}

\item{omega.se}{Numeric reflecting the standard errir for the
fourth path coefficient estimate (applicable only if T
rait = TRUE in \code{fit_SEM}). Defaults to NULL.}
}
\value{
A dataframe that contains four columns: a median of the
bootstrap values as an estimate of an indirect path coefficient,
lower and upper confidence intervals, and SE of the estimate.
}
\description{
\code{ind_path} estimates median and SE of an indirect pathway
by bootstrap
}
