% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionR2.R
\name{partitionR2}
\alias{partitionR2}
\title{Partition R2 in the full model}
\usage{
partitionR2(data = dat)
}
\arguments{
\item{data}{Data frame for a single study that contains all the
covariates needed to fit a single model}
}
\value{
A data frame with four columns, each for a variation explained by:
1) population size, 2)climate, 3) trait, 4) the full model.
}
\description{
\code{partitionR2} partitions the variation in the data explained by
fixed effects of the fitted GR-model (the model to explain population
growth rate). Currently only works for the model that includes
autocorrelation in residual structure and simpleSEM  option set to TRUE
}
\examples{
 # prepare the data to fit the model to one study only
 subs <- droplevels(dataSEM[dataSEM$ID == 7, ])
 full_NA <- data.frame(Year = seq(min(subs$Year), max(subs$Year), by = 1))
 consec_yrs <- merge(full_NA, subs, by = 'Year', all= TRUE)

 data_GR <- consec_yrs \%>\%
 dplyr::mutate(., Pop_mean_lag = c(Pop_mean[-1], NA)) \%>\%
 dplyr::mutate(., GR = log(Pop_mean_lag / Pop_mean)) \%>\%
 dplyr::filter(., !is.na(GR) & !is.na(Trait_mean) &
               !is.na(Demog_rate_mean) & !is.na(Pop_mean)) \%>\%
 dplyr::mutate(det_Clim = stats::resid(stats::lm(Clim ~ Year,
               data = .))) \%>\%
 dplyr::mutate(across(where(is.array), as.numeric))
 dat <- droplevels(data_GR[data_GR$ID == 7, ])

# partition the variation
partR2_study7 <- partitionR2(data = dat)

}
