% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloMat_descr.R
\docType{data}
\name{phyloMat}
\alias{phyloMat}
\title{phyloMat}
\format{
A matrix with 58 rows and 58 columns that contain
phylogenetic correlations between species
}
\usage{
phyloMat
}
\description{
Matrix of phylogenetic correlation between the
species for which phenological traits were measured
}
\keyword{datasets}
