% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_conc_DirInd.R
\name{pl_conc_DirInd}
\alias{pl_conc_DirInd}
\title{Plot of the direct effect of climate on population growth rate vs the
trait-mediated effect of climate on population growth rate}
\usage{
pl_conc_DirInd(
  Trait_categ = "Phenological",
  GlobES_dat = met_ef_T,
  ES_dat = wide_tempES,
  ClEfSpecific = TRUE
)
}
\arguments{
\item{Trait_categ}{Character specifying the level of the trait on which to subset the data.}

\item{GlobES_dat}{A data frame containing the across-study effect size estimates
per relation, obtained with the meta-analyses fitted using the function
\code{\link{fit_all_meta}}.}

\item{ES_dat}{A data frame containing the study-specific effect size estimates per
relation (in a wide format), extracted from prop_data object obtained with the
function \code{\link{fit_all_meta}}.}

\item{ClEfSpecific}{Boolean indicating whether the meta-analyses were fitted separately for
studies with positive and negative effect of climate on traits.}
}
\value{
A list with 1. a plot of study-specific estimates of direct effects
of climate on population growth rate (CG) vs the estimates of trait-mediated
effects of climate on population growth rate (CZG), overlayed with the across-study
effect sizes derived from the meta-analyses as well as the estimated relation between
CG and CZG using the meta-analysis; 2. correlation estimates between CZG and CG;
3. fitted mixed-effects model(s) predicting CG by CZG, if required per sign of the
climatic effect on traits.

## Still add examples
}
\description{
\code{pl_conc_DirInd} Plots for each study the direct effect of climate
on population growth rate vs the trait-mediated effect of climate on population
growth rate and overlays the across-study estimates obtained with meta-analyses
}
