% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concept.R
\name{plot_concept}
\alias{plot_concept}
\title{Plot of one chosen panel constituting conceptual plot (Z vs C, G vs Z|C or G vs C|Z)}
\usage{
plot_concept(
  Trait_categ = "Phenological",
  raw_dat = temp_std,
  GlobES_dat = met_ef_T,
  ES_dat = prop_T_simple,
  path = "CZ",
  xvar_raw = "det_Clim",
  yvar_raw = "Trait_mean",
  slope_ES = "Trait_mean<-det_Clim/Estimate",
  ylab = "Trait",
  xlab = "Climate"
)
}
\arguments{
\item{Trait_categ}{Character specifying the level of the trait on which to subset the data.}

\item{raw_dat}{A data frame with raw data that were used to fit single-study SEMs.
The data should be prepared in the way it was prepared for SEMs (e.g. climate, trait
and GR variables have to be standardized).}

\item{GlobES_dat}{A data frame containing the across-study effect size estimates
per relation, obtained with the meta-analyses fitted using the function
\code{\link{fit_all_meta}}.}

\item{ES_dat}{A data frame containing the study-specific effect size estimates per
relation (in a wide format), extracted from prop_data object obtained with the
function \code{\link{fit_all_meta}}.}

\item{path}{Character specifying which relation will be visualized on the plot.
Three possible: "CZ", "ZG" and "CG".}

\item{xvar_raw}{Character specifying what variable from the raw dataset (raw_dat)
plot on the x axis.}

\item{yvar_raw}{Character specifying what variable from the raw dataset (raw_dat)
plot on the y axis.}

\item{slope_ES}{Character specifying the name of the relation to be used to plot
from the ES_dat.}

\item{ylab}{Character specifying the label for the y axis.}

\item{xlab}{Character specifying the label for the x axis.}
}
\value{
Plots a requested relation (CZ, ZG or CG) for each study in the dataset and
overlays the across-study global effect sizes estimated with the meta-analyses.
## Still add examples + revise to get the xvar_raw, yvar_raw, and slopeES programmatically!!
}
\description{
\code{plot_concept} Plots one chosen relation from the path diagram:
Z vs C, G vs Z|C or G vs C|Z for each specific study and adds the
across-study effect sizes estimated with meta-analytical models
}
