% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest plot of effect sizes per each study and global effect sizes across studies}
\usage{
plot_forest(
  data_ES,
  data_globES,
  Cov_fact = NULL,
  COV = NULL,
  xlab = "Effect of temperature on trait",
  colr = c("blue", "green", "black"),
  pdf_basename = NULL,
  mar = c(4, 10, 2, 2),
  labels_ES = TRUE
)
}
\arguments{
\item{data_ES}{Data frame containing, for each study, the effect size estimates and their
standard errors, returned as 'data_EfS' by the function \code{\link{fit_meta_phylo}}.}

\item{data_globES}{Data frame containing the global effect sizes across the studies,
returned by function \code{\link{fit_meta_phylo}} as either 'data_Covar' or 'data' (depending on
whether the explanatory categorical variable was included in the meta-analytical model).}

\item{Cov_fact}{A character specifying the name of the categorical variable to be included
as a fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case no
categorical variables are included and the overall global effect size is estimated.}

\item{COV}{A character specifying the name (or names, separated by '+') of the continuous
variable to be included as a fixed-effect covariate in the meta-analysis. Defaults to NULl,
in which case no continuous covariates are included and the overall global effect size
is estimated.}

\item{xlab}{Character specifying the label for the x axis.}

\item{colr}{Vector specifying the colours to be used for the data points. The length of the
vector should correspond to the number of the levels in the categorical explanatory variable
included in the meta-analytical model, or should be one (if the single global effect size
across all studies is to be plotted).}

\item{pdf_basename}{A character specifying the name of the .pdf for
the produced plot.}

\item{mar}{A vector specifying the plot margins, analogously to \code{\link[graphics]{par}}.}

\item{labels_ES}{Boolean specifying whether to plot the labels for all the single studies.
Defaults to TRUE.}
}
\value{
Plots a forest plot with effect sizes (and SEs) for each study,
and global effect size(s) on the bottom. The data used for the plot are
returned (invisibly).
}
\description{
\code{plot_forest} Plots a forest plot displaying the effect sizes
for each study and global effects across the studies
}
