% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_points.R
\name{plot_hist_points}
\alias{plot_hist_points}
\title{Plot of the across-study effect sizes on top of the histogram
of all the effect sizes}
\usage{
plot_hist_points(
  data_allEstim,
  Ef_sizesEstim,
  dataAxis,
  Traitdem = "Morphology + Recruitment",
  tabC,
  angle_yax_lab = 90,
  annot = NULL
)
}
\arguments{
\item{data_allEstim}{A data frame with effect size estimates per relation for each
study (as returned by the SEM fitted with the function \code{\link{fit_SEM}}.}

\item{Ef_sizesEstim}{A data frame containing the across-study effect size estimates
per relation, obtained with the meta-analyses fitted using the function
\code{\link{fit_all_meta}}.}

\item{dataAxis}{A data frame specifying the names of the relations to be plotted on the
y axis.}

\item{Traitdem}{Character specifying the level of the trait, for which analyses were
conducted.}

\item{tabC}{A data frame containing the number of studies per each combination of
trait and demographic rate category.}

\item{angle_yax_lab}{Numeric specifying whether the labels of relations provided with
\code{dataAxis} are to be plotted parallel (\code{dataAxis} = 0) or perpendicular
(\code{dataAxis} = 90) to the x axis.}

\item{annot}{A data frame supplying the information on whether each relation was
estimated with complex or simple SEM (see \code{\link{fit_mod}} for more details),
so that the respective plot panesl will be shaded accordingly. Applicable only if
the plot combines the relations from simple and complex SEMs, otherwise set to NULL
(default).}
}
\value{
Plots a forest-like plot with, in addition to the overall effect size estimates
shown by points and error bars, shown historgams of all the underlying study-speciifc
effect sizes. Each panel corresponds to a single relation (see \code{Type_EfS} in
\code{\link{fit_meta}} function) and effect sizes for precipitation and temperature
are shown in different colors and shapes.
## Still add examples
}
\description{
\code{plot_hist_points} Plots across-studies effect sizes as points and error bars on
top of the histogram of the underlying study-species effect sizes, per combination of
trait and demographic rate category
}
