% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lab_name.R
\name{plot_lab_name}
\alias{plot_lab_name}
\title{Providing labels for axes in forest plot and 'relation plots', as well as names for the pdfs}
\usage{
plot_lab_name(
  Relation = "GR<-det_Clim",
  Cov_fact = NULL,
  Demog_rate = "Survival",
  Trait_categ = "Phenological",
  Clim = "Temperature"
)
}
\arguments{
\item{Relation}{Character specifying for which type of the effect size a meta-analysis
was conducted. Possible are: 'Demog_rate_mean<-det_Clim', 'Demog_rate_mean<-Pop_mean',
'Demog_rate_mean<-Trait_mean', 'GR<-Demog_rate_mean', 'GR<-det_Clim', 'GR<-Pop_mean',
'Ind_DemRate<-det_Clim', 'Ind_GR<-det_Clim', 'Tot_DemRate<-det_Clim',
'Tot_GR<-det_Clim', and 'Trait_mean<-det_Clim'. For more details see
\code{\link{fit_meta_phylo}}.}

\item{Cov_fact}{A character specifying the name of the categorical variable to be included
as a fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case no
categorical variables are included and the overall global effect size is estimated.}

\item{Demog_rate}{Character specifying the level of the demographic rate for which
analyses were conducted.}

\item{Trait_categ}{Character specifying the level of the trait, for which analyses were
conducted.}

\item{Clim}{Character specifying the level of the climatic variable, for which analyses
were conducted.}
}
\value{
A list with an x axis lable for the forest plot, x- and y-axes labels for 'relation plots'
and a nem for the pdf.
}
\description{
\code{plot_lab_name} Prepares axes labels for forest plots and 'relation plots', as well as
names for the pdfs with plots
}
\examples{
lab_name <- plot_lab_name(Relation = 'GR<-det_Clim',
                          Cov_fact = NULL,
                          Demog_rate = 'survival',
                          Trait_categ = 'phenological',
                          Clim = 'temperature')

}
