% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uni_spSpec.R
\name{plot_uni_spSpec}
\alias{plot_uni_spSpec}
\title{Plot predictions of climate effects on traits, using species-specific predictors}
\usage{
plot_uni_spSpec(
  data_allEstim = CZ_Phen,
  mod_mv = mod_CZ_Lat,
  lOut = 10,
  pdf_basename = NULL,
  xLab
)
}
\arguments{
\item{data_allEstim}{A data frame with effect size estimates per relation for each
study (as returned by the SEM fitted with the function \code{\link{fit_SEM}}.}

\item{mod_mv}{Model object of class rma.mv fitted to test the
hypothesis about the effect of the specified species characteristic..}

\item{lOut}{Numeric specifying the length of the predictor variable, for which the
response of species traits to climate will be predicted and plotted.}

\item{pdf_basename}{A character specifying the name of the .pdf for
the produced plot.}

\item{xLab}{Character specifying the x axis label.}
}
\value{
Plots a forest plot with effect sizes (and SEs) for each study,
and global effect size(s) on the bottom. The data used for the plot are
returned (invisibly).
}
\description{
\code{plot_uni_spSpec} Plots predictions of the model using as a response effect
of climate on trait and as predictor one species-specific predictor at a time and
climate
}
\examples{
STILL To add

}
