% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SEM_input.R
\name{prep_SEM_input}
\alias{prep_SEM_input}
\title{Prepare a dataset for SEM based on the output of window analyses}
\usage{
prep_SEM_input(prep_subset_climwin, oneGrid = FALSE, explanYear = TRUE,
  endWindow = 104, RefMon = NA,
  selIDs = unique(prep_subset_climwin$Sel[[1]]$ID))
}
\arguments{
\item{prep_subset_climwin}{The object returned by \code{\link{prep_subset}}
and that was used for window analyses.}

\item{oneGrid}{Logical (TRUE/FALSE). Whether climatic data for window analyses
were extracted from a single grid cell into which the study location falls or
if the mean of five cells (the focal one and four neighbours) was used.}

\item{explanYear}{Logical (TRUE/FALSE). Whether year was included as an
explanatory variable in the baseline formula for window analysis.}

\item{endWindow}{A numeric specifying the furtherst number of time intervals
(set by cinterval) back from the reference day. See also the option range in
the function \code{\link[climwin]{slidingwin}}.}

\item{RefMon}{A numeric specifying the month in case the absolute window was
for all species. If NA was specified then a species-specific window was used for
each dataset based on the available data on phenology and the period when morphology
was recorded.}

\item{selIDs}{A vector specifying the study IDs to be selected to include in the
dataset for SEM analyses.}
}
\value{
Returns a dataframe that for each study ID provides the data needed to conduct
SEM, i.e. climatic data extracted with window analyses, mean yearly trait values (with
SEs if available), mean yearly demographic rate values (with SEs if available) and
yearly population size estimates (with SEs if available). For each study ID metadata (e.g.
the study species, location, coordinates) are also included.
}
\description{
\code{prep_SEM_input} produces a dataset to be used for SEM based on
the results of window analyses and supplied study IDs
}
\examples{
temp_eu_SEM <- prep_SEM_input(prep_subset_climwin = eu_noSea,
oneGrid = FALSE,
explanYear = TRUE,
endWindow = 104,
RefMon = NA,
selIDs = 152)
}
