% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_subset.R
\name{prep_subset}
\alias{prep_subset}
\title{Prepare a subset of data focusing on a particular group of species}
\usage{
prep_subset(data, Seabird = FALSE)
}
\arguments{
\item{data}{Data frame to use for subsetting.}

\item{Seabird}{Logical (TRUE/FALSE) specifying whether to include
sea birds in the subset or not.}
}
\value{
Returns a tibble with three columns: "Sel" - a data frame where
each row represents a unique record per study ID for which climwin analyses
have to be conducted; "biol_NY" - a data frame with all the information as in
original study (longitudinal data per each study ID) and, additionally, NYears
column specifying the number of years per study; "subdata" -
a subset of biological data to be used in climwin analyses.
}
\description{
\code{prep_subset} produces a subset of the supplied data
according to the requirements, i.e. whether the subset should be
for sea birds or not
}
\examples{
biol_dat <- read.csv('./data-raw/Data_11_06_2019.csv')
biol_noSea <- prep_subset(data = biol_dat, Seabird = FALSE)
}
