% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_path.R
\name{prop_path}
\alias{prop_path}
\title{Obtain proportional contributions of each pathway to the total pathway}
\usage{
prop_path(data, data_MA, DD = "n_effectDGR")
}
\arguments{
\item{data}{Data frame containing, for each study, the effect size estimates of all
pathways, also combined ones. This dataframe is returned as an intermediate step of
the function \code{\link{fit_meta}}.}

\item{data_MA}{Data frame containing, for each study, the effect size estimates
for each pathway from the SEM analyses. This data frame also contains meda-data
(e.g. sutdy species, study location, continent, life history traits of the species),
that are needed to fit the mixed-effect model. See also \code{\link{fit_meta}}}
}
\value{
Returns a data frame, that in addition to the meta-data for each study, contains
proportional contributions of each pathway (direct, indirect, and others, if present) to
the total pathway from climate to demographic rate and from climate to the population growth
rate.
}
\description{
\code{prop_path} estimates proportional contributions of each path (direct,
indirect and others if present) to the two total pathways: one from climate to
demographic rate and the other one from climate to population growth rate. This
function is called from within a function \code{\link{fit_meta}}.
}
\examples{
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
check_noCovar <- fit_meta(data_MA = Coefs_Aut, Type_EfS = 'Trait_mean<-det_Clim',
                          Covar = NULL)
check_TraitCateg <- fit_meta(data_MA = Coefs_Aut, Type_EfS = 'Trait_mean<-det_Clim',
                             Covar = 'Trait_Categ')
check_TraitCateg
}
