% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_stats}
\alias{replace_stats}
\title{Replace the stats in the table by the correct values from the omnibus test}
\usage{
replace_stats(data, variable, stats_out)
}
\arguments{
\item{data}{A dataframe containing the output of the model summary}

\item{variable}{A character specifying the name of the variable for which the omnibus
test is ot be computed.}
}
\value{
A dataframe that contains corrected statistics values for the specified variable.
}
\description{
\code{replace_stats} replaces the values in the table by the correct output of
omnibus test.
}
