% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xlsx.R
\name{save_xlsx}
\alias{save_xlsx}
\title{Save a table in the *.xlsx format}
\usage{
save_xlsx(table, table_name, first_column_name = NULL)
}
\arguments{
\item{table}{A data frame oject to be saved as .xslx.}

\item{table_name}{A string specifying the path and the name of the table to be created.}

\item{first_column_name}{A string indicating the name of the first column which is taken from
the row names of the table (DROP???)}
}
\value{
The table (but the return is invisible).
}
\description{
This function is called internally by other functions. It is a wrapper for
the function \code{\link[openxlsx:write.xlsx]{write.xlsx}} of the package \pkg{openxlsx}.
}
\examples{
d <- data.frame(id = LETTERS[1:5], test = 1:5,
                a= seq(from = 0.01, to = 0.05, by = 0.01),
                 b = c(1,2 , 0.03, 0.004, 0.1),
                p = c('<0.001', '0.002', '0.260', '<0.001', '0.987'), test = 1:5)
dir.create('tables')
save_xlsx(table = d, table_name = './tables/test1')
unlink('tables', recursive=TRUE)

}
