% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_CorClimEff.R
\name{tab_CorClimEff}
\alias{tab_CorClimEff}
\title{Save a table with statistics for correlations among the climate effects on traits}
\usage{
tab_CorClimEff(data, table_name, explanatory)
}
\arguments{
\item{data}{Data frame including for each study the correlation between
the effects of precipitaion and temperature on traits, and explanatory variables.
The correlation has to be coded as binary (1/0), where 1 = positive, and 0 =
negative correlation.}

\item{table_name}{A string specifying the path and the name of the
table to be created.}

\item{explanatory}{Character indicating the name of the variable whose
impact on correlation between the effects of precipitation and temperature on
traits is to be tested.}
}
\value{
A list containing two objects: (1) a dataframe with the statistics for
the effect of a specified predictor on the correlation between the effects of precipitation
and temperature on traits; and (2) a model object for the full model. The returned table with
statistics is saved (invisibly) as an .xlsx sheet.
}
\description{
\code{tab_CorClimEff} saves the table with the statistics extracted
from the model fitted to assess the effect of specific predictors on
the correlation between the temperature and precipitation effects on traits.
}
\examples{
Coef_ClTr_TP <- readRDS(file = './data/Corr_ClTrEffects.rds')
lat_m <- tab_CorClimEff(data = Coef_ClTr_TP,
                        table_name = './tables/Latitude_Eff_CorClimEffects',
                        explanatory = 'Latitude')

}
