% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_spSpecific.R
\name{tab_spSpecific}
\alias{tab_spSpecific}
\title{Save a table with statistics for species-specific analyses}
\usage{
tab_spSpecific(mod_mv, table_name, explanatory)
}
\arguments{
\item{mod_mv}{Model object of class rma.mv fitted to test the
hypothesis about the effect of the specified species characteristic
on the indirect path.}

\item{table_name}{A string specifying the path and the name of the
table to be created.}

\item{explanatory}{Character indicating the name of the variable
corresponding to a species characteristic, for which the
hypothesis is being tested.}
}
\value{
A dataframe with the statistics for the test of the effect of
a specified characteristic on the indirect path, and saves this dataframe
(invisibly) as an .xlsx sheet.
}
\description{
\code{tab_spSpecific} saves the table with the statistics exracted
from the model fitted to test the hypothesis about the effect of
a specified species characteristic on the indirect path.
}
\examples{
Coefs_Aut <- readRDS(file = './output_overall_sig/PathCoefs_allMods_Temp_AlsoEstimatedRelations.RDS')
traits <- read.csv('./data-raw/Species_traits_Subset_11_09.csv')
Coef_Aut_IndDR <- droplevels(Coefs_Aut \%>\%
dplyr::filter(., Relation == 'Ind_DemRate<-det_Clim'))
Coefs_Aut_sp <- base::merge(Coef_Aut_IndDR, traits_proc, by = 'Species')
mod_genLength <- metafor::rma.mv(Estimate ~ GenLength_y_IUCN + Trait_Categ + Demog_rate_Categ +
                            GenLength_y_IUCN * Trait_Categ + GenLength_y_IUCN * Demog_rate_Categ +
                            Pvalue,
                            V = SError^2, random = list(~ 1|Species, ~1|ID, ~1|Location),
                            data = Coefs_Aut_sp, method = 'ML')
tab_spSpecific(mod_mv = mod_genLength, table_name = './tables/GenLength_Temp', explanatory = 'GenLength_y_IUCN')

}
