% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_spSpecific_uni.R
\name{tab_spSpecific_uni}
\alias{tab_spSpecific_uni}
\title{Save a table with statistics for species-specific analyses}
\usage{
tab_spSpecific_uni(mod_mv, table_name, explanatory, interact_fac = NULL)
}
\arguments{
\item{mod_mv}{Model object of class rma.mv fitted to test the
hypothesis about the effects of the specified species characteristic(s)
on a certain path from SEMs.}

\item{table_name}{A string specifying the path and the name of the
table to be created.}

\item{explanatory}{A character string indicating the name(s) of
the variable(s) corresponding to (a) species characteristic(s),
for which the hypothesis is being tested.}

\item{interact_fac}{A character specifying the name of the factor
variable with which interactions of other variable(s) are tested,
defaults to NULL.}
}
\value{
A dataframe with the statistics for the test of the effect(s) of
(a) specified characteristic(s) on the specific path from SEMs,
and saves this dataframe (invisibly) as an .xlsx sheet.
}
\description{
\code{tab_spSpecific_uni} saves the table with the statistics extracted
from the model fitted to test hypothesis about the effects of
specified species characteristic(s) on the path from SEMs.
}
\examples{
dataPaths_phen <- dataPaths \%>\%
    dplyr::filter(Trait_Categ == 'Phenological') \%>\%
    dplyr::mutate(absLat = abs(Latitude))

mod_CZ_PhenT_AbsLat <- metafor::rma.mv(Estimate ~ absLat + Pvalue,
                              V = Std.Error^2, random = list(~ 1|Species, ~1|ID, ~1|Location),
                              data = dataPaths_phen, method = 'ML')
summary(mod_CZ_PhenT_AbsLat)
tab_spSpecific_uni(mod_mv = mod_CZ_PhenT_AbsLat,
                   table_name = './tables/CZ_PhenT_AbsLat',
                   explanatory = c('absLat', 'Pvalue'),
                   interact_fac = NULL)

}
