% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_paths.R
\name{tot_path}
\alias{tot_path}
\title{Provide estimates of median and SE of the total pathway by bootstrap}
\usage{
tot_path(
  direct,
  indir,
  ClDem = NULL,
  DemGR = NULL,
  direct.se,
  indir.se,
  ClDem.se = NULL,
  DemGR.se = NULL,
  numrep = 10000
)
}
\arguments{
\item{direct}{Numeric reflecting the estimate of the direct
path coefficient.}

\item{indir}{Numeric reflecting the estimate of the indirect
path coefficient.}

\item{ClDem}{Numeric reflecting the estimate of the path from
climate to demographic rate, in case the total path from
climate to growth rate is estimated. Defaults to NULL.}

\item{DemGR}{Numeric reflecting the estimate of the path from
demographic rate to growth rate, in case the total path from
climate to growth rate is estimated. Defaults to NULL.}

\item{direct.se}{Numeric reflecting the standard error for the direct
path coefficient.}

\item{indir.se}{Numeric reflecting the standard error for the indirect
path coefficient.}

\item{ClDem.se}{Numeric reflecting the standard error for the path from
climate to demographic rate.}

\item{DemGR.se}{Numeric reflecting the standard error for the path from
demographic rate to growth rate.}

\item{numrep}{Numeric for the number of replicates to use in a
bootstrap.}
}
\value{
A dataframe that contains four columns: a median of the
bootstrap values as an estimate of an indirect path coefficient,
lower and upper confidence intervals, and SE of the estimate.
}
\description{
\code{tot_path} estimates median and SE of the total pathway
by bootstrap
}
