% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_combi_paths.R
\name{all_combi_paths}
\alias{all_combi_paths}
\title{Use bootstrap to obtain the coefficients and SEs for all combined pathways}
\usage{
all_combi_paths(
  data_forMA,
  DD = "n_effectDGR",
  simpleSEM = FALSE,
  Trait = FALSE
)
}
\arguments{
\item{DD}{Character specifying how to account for density dependence by
including population size as additional explanatory in the SEM. Possibilities
are: 'none' - no inclusion of density dependence, 'n_effectGR' - effect
of population size on growth rate, 'n_effectD' - effect of population size
on demographic rate, 'n_effectDGR' - effect of population size on demographic
rate and growth rate. Only options 'n_effectGR' and 'none' are possible for
simple structure of SEM, in which climate effect on GR i sonly mediated by trait
(i.e. \code{simpleSEM = TRUE}).}

\item{simpleSEM}{Logical (TRUE/FALSE) specifying that the simple SEM
structure will be fitted in which climate effect on GR is only mediated
by trait, demographic rates are not considered. Defaults to FALSE.}

\item{Trait}{Logical (TRUE/FALSE) specifying whether to include
the direct impact of trait on GR in the SEM fitted using a complex
structure, i.e. climate effect on GR is mediated by both trait and
demographic rate. Is applicable only if \code{simpleSEM = FALSE}.
Defaults to FALSE.}

\item{data_for_MA}{Data frame containing a subset of data_MA (see
\code{fit_meta_phylo}) in a wide format, i.e. single column per pathway.
This data frame contains for each pathway for each study the
effect size estimates along with their SEs obtained with the SEM
analyses. Besides that, some meta-data on each study is included
(e.g. study species, study location, study ID).}
}
\value{
Returns a data frame, which is an extended version of the input data frame,
that additionally contains estimated medians, SEs, as well as lower and upper CI
for the combined pathways.
}
\description{
\code{all_combi_paths} returns the medians and their SE estimates for
all combined pathways present in SEM, using the bootstrap procedure
}
\examples{
prepare the data to fit the model to extract indirect path
dataPaths_sp <- dataPaths \%>\%
                  dplyr::mutate(Species = dplyr::case_when(
                         Species == 'Cyanistes caeruleus' ~ 'Parus caeruleus',
                         Species == 'Thalasseus sandvicensis' ~ 'Sterna sandvicensis',
                         Species == 'Setophaga caerulescens' ~ 'Dendroica caerulescens',
                         Species == 'Thalassarche melanophris' ~ 'Thalassarche melanophrys',
                         Species == 'Ichthyaetus audouinii' ~ 'Larus audouinii',
                         Species == 'Stercorarius maccormicki' ~ 'Catharacta maccormicki',
                         TRUE ~ Species))

dataPaths_sp$Species <- unlist(lapply(1:nrow(dataPaths_sp), FUN = function(x){
  binary <- strsplit(as.character(dataPaths_sp$Species[x]), " ")
  Underscore <- paste(binary[[1]][1], binary[[1]][2], sep = "_")}))
dataPaths_sp$Sp_phylo <- dataPaths_sp$Species

# fit the models for: Trait_mean<-det_Clim', 'Ind_GR<-det_Clim', 'Tot_GR<-det_Clim'
meta_Phen_Cov <- fit_all_meta(data_MA = dataPaths_sp,
                              Demog_rate = NULL,
                              Trait_categ = 'Phenological',
                              Clim = 'Temperature',
                              Cov_fact = 'WeathQ',
                              COV = 'Pvalue',
                              sel = 'Temp_Phen_Cov',
                              folder_name = NULL,
                              colr = c('black', 'red'),
                              DD = 'n_effectGR',
                              simpleSEM = TRUE,
                              A = phyloMat,
                              all_Relations = c('Trait_mean<-det_Clim',
                              'Ind_GR<-det_Clim', 'Tot_GR<-det_Clim'))
two <- dplyr::bind_rows(lapply(list('Ind_GR<-det_Clim', 'Tot_GR<-det_Clim'),
function(x){extr_coefs(obj = meta_Phen_Cov, Type_EfS = x)}))
# fixing inconsistent variable naming
Coefs_subst <- dataPaths_sp \%>\%
dplyr::rename(.,  SError = Std.Error)
two$P.Value = rep(NA, nrow(two))

Coefs_subst <- Coefs_subst \%>\%
    dplyr::select(., -c(names(Coefs_subst)[! names(Coefs_subst) \%in\% names(two)]))
two <- two \%>\%
    dplyr::select(., -c(names(two)[! names(two) \%in\% names(Coefs_subst)]))
Coef_all <- rbind(Coefs_subst, two) \%>\%
dplyr::filter(Trait_Categ == 'Phenological')
 forTrans <- subset(Coef_all, select = c(Estimate,  SError, Relation,
 Species, Location, ID))
 met_wide <- forTrans \%>\%
     tidyr::gather(variable, value, -(Relation:ID)) \%>\%
     tidyr::unite(temp, Relation, variable, sep = '/') \%>\%
     tidyr::spread(temp, value)
 test_simSEM <- all_combi_paths(data = met_wide, DD = 'n_effectGR',
                                simpleSEM = TRUE, Trait = FALSE)

}
