% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_climwin.R
\name{analyse_climwin}
\alias{analyse_climwin}
\title{Analyze the results of climate window analysis}
\usage{
analyse_climwin(
  ID,
  biol_data,
  out_clim = "output_climwin_temp",
  randwin = FALSE,
  metric = "AIC",
  MinDur = 1,
  MaxDur = 40,
  deltaThresh = -7,
  test_winDur = FALSE,
  out_for_SEM = "output_SEM_test",
  oneGrid = TRUE,
  explanYear = TRUE,
  endWindow,
  RefMon = NA
)
}
\arguments{
\item{ID}{Numeric giving a unique ID of the current dataset for
a given population and species.}

\item{biol_data}{Data frame with trait data for a given
population and species.}

\item{out_clim}{Character specifying the library on the path where
the results of climate window analysis were stored.}

\item{randwin}{Logical (TRUE/FALSE) specifying whether \code{\link[climwin]{randwin}}
was run together with \code{\link[climwin]{slidingwin}}.}

\item{metric}{Character specifying 'AIC' or 'C'. This define whether
a value of PDAIC or Pc will be returned. For more information
see the same option in the function \code{\link[climwin]{pvalue}}.}

\item{MinDur}{A numeric specifying the minimum allowed window duration.}

\item{MaxDur}{A numeric specifying the maximum allowed window duration.}

\item{deltaThresh}{A numeric specifying the minimum delta AIC with the null
model to accept a specific climatic window model.}

\item{test_winDur}{A Boolean specifying whether to check if the identified
window lays within the specified range and the delta AIC value is larger than
the specified threshold.}

\item{out_for_SEM}{Character specifying the library on the path where
to store the results that are to be used as input for SEM.}

\item{oneGrid}{Logical (TRUE/FALSE). Whether to extact climatic data from
a single grid cell where into which the study location falls or to use mean of
five cells: the focal one and four neighbours.}

\item{explanYear}{Logical (TRUE/FALSE). Whether to include year as an
explanatory variable in the baseline formula for window analysis.}

\item{endWindow}{A numeric specifying the furthest number of time intervals
(set by cinterval) back from the reference day. See also the option range in
the function \code{\link[climwin]{slidingwin}}.}

\item{RefMon}{A numeric specifying the month for the absolute window
to be used for all species. If NA is specified then a species-specific window is used for
each dataset using the available data on phenology and the period when morphology
was recorded.}
}
\value{
If the results are not likely to be an issue of overfitting,
then the function returns a tibble with four columns: "ID"
- study id, "Species" - study species, "pvalue" - the pvalue from
the test assessing whether the climate signal is obtained by chance only,
and "data_res" - a data frame with all the information required
for running the SEM. If the results are likely an issue of
overfitting, then the returned tibble contains only first three of
the above-mentioned columns.
}
\description{
\code{analyse_climwin} analyses the results obtained
with climate window analysis (function \code{\link{climwin_proc}})
for the specified study ID
}
\examples{
# ATTENTION: DO NOT RUN! takes long time
\dontrun{
biol_noSea <- prep_subset(data = data_biol, Seabird = FALSE)
# keep only EU countries
biol_eu <- droplevels(subset(biol_noSea$subdata[[1]],
! Country \%in\% c('Antarctica', 'Australia',
                'Canada', 'Falkland Islands',
                'Greenland', 'Mexico',
                'New Zealand', 'South Africa',
                'South Atlantic Ocean',
                'South Georgia', 'Svalbard',
                'Taiwan', 'USA', 'Venezuela')))
meanT <- raster::stack(x = system.file("extdata",
"tg_ens_mean_0.1deg_reg_v18.0e.nc", package="sTraitChange"))
test_rand <- climwin_proc(biol_data = biol_eu,
                          clim_data = meanT, ID = 1,
                          randwin = FALSE, seednum = 1302,
                          repeats = 30, plot_check = FALSE,
                          out_clim = tempdir(), # attention: for this example we
                                     # write the data to a temporary directory,
                                     # to check its location type tempdir()
                          cinterval = 'month',
                          stat = 'mean',
                          startWindow = 0, endWindow = 12,
                          oneGrid = FALSE, explanYear = TRUE,
                          RefMon = NA, weatherVar = NA)
message('Temporary directory is located at', tempdir())

                t_anal <- analyse_climwin(ID = 1,
                biol_data = biol_eu,
                out_clim = tempdir(), # attention: for this example we
                                     # wrote the output of climwin_proc
                                     # to a temporary directory,
                                     # so also have to get it from there
                randwin = FALSE, metric = 'AIC',
                MinDur = 1, MaxDur = 40,
                deltaThresh = -7, test_winDur = FALSE,
                out_for_SEM = tempdir(), # attention: for this example we
                                     # again write the data to a temporary directory,
                                     # but the user may decide to specify a different
                                     # directory
                oneGrid = FALSE, explanYear = TRUE,
                endWindow = 12, RefMon = NA)}
message('Temporary directory is located at', tempdir())
}
