% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climwin_proc.R
\name{climwin_proc}
\alias{climwin_proc}
\title{Conduct climate window analysis}
\usage{
climwin_proc(biol_data, clim_data, ID, randwin = FALSE, seednum = 1302,
  repeats = 20, plot_check = FALSE, RefMon = 6,
  out_dir = "output_climwin", stat = "mean", region = "Europe")
}
\arguments{
\item{biol_data}{Data frame with trait data for a given population and species.}

\item{clim_data}{Raster stack of climatic data across Europe
(if the study is outside Europe use XX).}

\item{ID}{Numeric giving a unique ID of the current dataset for a given population and species.}

\item{randwin}{Logical (TRUE/FALSE). Should \code{\link[climwin]{randwin}}
be run together with \code{\link[climwin]{slidingwin}}?}

\item{repeats}{If randwin is run, the number of times that data will be randomised.}

\item{plot_check}{Logical (TRUE/FALSE). Whether to display
one year (the first in the series) of the climatic data
 together with the study location.}

\item{RefMon}{Numeric specifying the month for the refday(),
see the same option in the function \code{\link[climwin]{slidingwin}}
for more information.}

\item{out_dir}{Character specifying the library on the path where
to save the results}

\item{stat}{Character specifying which statistics to use for
aggregating the climatic data (see also the same option in
\code{\link[climwin]{slidingwin}}).}

\item{region}{Character specifying from which region the climatic data is
extracted. For now supported are: 'Europe' and 'USA'. Defaults to 'Europe'.}
}
\value{
A tibble with six columns if randwin is set to FALSE
and seven columns if randwin is set TRUE. The columns are: "ID"
- study id, "Species" - study species, "Trait" - the trait considered,
"climwin_output" - a list returned by slidingwin(),
"randwin_output" - a list returned by randwin (if set to TRUE),
"clim_data" - a data frame with climate values,
"biol_data" - a data frame with traits, demographic rates
and population size.
}
\description{
\code{climwin_proc} performs climate window analysis for
specified traits and climate data and returns the best
identified window together with results of randomisation
tests (if asked for)
}
\examples{
dat_birds <- read.csv('./data-raw/Test_european_birds.csv')
meanT <- raster::stack('./data-raw/tg_ens_mean_0.1deg_reg_v18.0e.nc')
test_rand <- climwin_proc(biol_data = dat_birds,
                          clim_data = meanT, ID = 1,
                          randwin = FALSE, seednum = 1302,
                          repeats = 30, plot_check = FALSE,
                          RefMon = 6, out_dir = 'output_climwin',
                          stat = 'mean')

}
