% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_res_SEM.R
\name{extract_res_SEM}
\alias{extract_res_SEM}
\title{Extract results of the fitted SEM concerning a certain aspect of the model}
\usage{
extract_res_SEM(list_fitSEM, stat_extr)
}
\arguments{
\item{list_fitSEM}{List of SEMs fitted with the funciton \code{\link{fit_SEM}}.}

\item{stat_extr}{Character specifying on which aspect of the fitted SEM
you want to extract the results. Can take on the following values: 'coefs',
'R2', 'Cstat'.}
}
\value{
A dataframe containing the meta-data on each study (e.g. study is,
species name, location, trait category, demographic rate category etc) and
the data on the requested aspect of the fitted SEM.

still add the example
}
\description{
\code{extract_res_SEM} extracts the results of the list of models fitted with
\code{\link{fit_SEM}} function. The results are extracted with a focus on a particular aspect
of the model: Cstat value, R2 values or obtained path coefficients
}
