% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_meta.R
\name{fit_all_meta}
\alias{fit_all_meta}
\title{Fit all meta-analytical models for each relation in the SEM,
for a specified subselection of demographic rate and trait}
\usage{
fit_all_meta(
  data_MA,
  Demog_rate = "Survival",
  Trait_categ = "Phenological",
  Clim = "Temperature",
  tab = "Taxon",
  Covar = NULL,
  sel = "Phen_Surv",
  folder_name = NULL,
  colr = c("black")
)
}
\arguments{
\item{data_MA}{Data frame containing, for each study, the effect size estimates
for each pathway from the SEM analyses. This data frame also contains meda-data
(e.g. sutdy species, study location, continent, life history traits of the species),
that are needed to fit the mixed-effect model.}

\item{Demog_rate}{Character specifying the level of the demographic rate on which
to subset the data.}

\item{Trait_categ}{Character specifying the level of the trait on which to subset the data.}

\item{Clim}{Character specifying the level of the climatic variable on which to subset
the data.}

\item{tab}{Categorical specifying the name of the categorical variable for which
to produce the frequency table.}

\item{Covar}{Categorical specifying the name of the categorical variable to be included
as fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case no
categorical variables are included and the overall global effect size is estimated.}

\item{sel}{Character specifying the name to be included in the .pdf name, which indicates
the levels of demographic rate and trait for which the subsetting was done.}

\item{folder_name}{Character specifyng the path to the directory in which the results will be saved.}

\item{colr}{Vector specifying the colours to be used for the data points. The length of the
vector should correspond to the number of the levels in the categorical explanatory variable
included in the meta-analytical model, or should be one (if the single global effect size
across all studies is to be plotted).}
}
\value{
Returns a tibble that includes one data frame, one tibble and a table. A data frame
contains the estimated effect sizes, their standard errors, their significance and the AIC per each
fitted mixed-effects model, including also the column 'Relation' specifying for which relation
the data were analyzed (e.g. Demog_rate_mean<-Pop_mean', for more details see \code{\link{fit_meta}}).
A tibble contains two lists and one character variable. The first list contains estimated global
effect sizes, their standard errors, their significance and the AIC per each fitted mixed-effects model
defined by the 'Relation', as returned by \code{\link{fit_meta}}. The second list contains data frames
with the effect sizes and standard errors for each study. Each data frame corresponds to the subset
of the data per each 'Relation' type. A character 'names' specifies the 'Relation' type. A table
contains frequencies for the levels of the specified categorical variable, supplied via parameter 'tab'.
}
\description{
\code{fit_all_meta} fits all mixed-effects meta-analytical models (per each relation
in the SEM), to extract global effect sizes across the studies in a specified subset
of data as defined by a demographic rate and a trait
}
\examples{
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
meta_Phen_Surv <- fit_all_meta(data_MA = Coefs_Aut,
                               Demog_rate = 'survival',
                               Trait_categ = 'phenological',
                               Clim = 'temperature',
                               tab = 'Taxon',
                               Covar = NULL,
                               sel = 'Phen_Surv',
                               folder_name = './output_overall/',
                               colr = c('black'))
meta_Phen_Surv
meta_Phen_Surv_byCont <- fit_all_meta(data_MA = Coefs_Aut,
                                      Demog_rate = 'survival',
                                      Trait_categ = 'phenological',
                                      Clim = 'temperature',
                                      tab = 'Continent',
                                      Covar = 'Continent',
                                      sel = 'Phen_Surv',
                                      folder_name = './output_overall/',
                                      colr = c('black', 'green', 'blue', 'red'))
}
