% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_meta.R
\name{fit_all_meta}
\alias{fit_all_meta}
\title{Fit all meta-analytical models for each relation in the SEM,
for a specified subselection of demographic rate and trait}
\usage{
fit_all_meta(
  data_MA,
  Demog_rate = "Survival",
  Trait_categ = "Phenological",
  Clim = "Temperature",
  Cov_fact = NULL,
  COV = NULL,
  sel = "Phen_Surv",
  folder_name = NULL,
  colr = c("black"),
  DD = "n_effectDGR",
  simpleSEM = FALSE,
  Trait = FALSE,
  A = Mat_phylo,
  des.matrix = "treatm.contrasts",
  all_Relations = c("Demog_rate_mean<-det_Clim", "Demog_rate_mean<-Pop_mean",
    "Demog_rate_mean<-Trait_mean", "GR<-Demog_rate_mean", "GR<-det_Clim", "GR<-Pop_mean",
    "Ind_DemRate<-det_Clim", "Ind_GR<-det_Clim", "Tot_DemRate<-det_Clim",
    "Tot_GR<-det_Clim", "Trait_mean<-det_Clim", "Ind_GR<-Pop_mean", "Tot_GR<-Pop_mean")
)
}
\arguments{
\item{data_MA}{Data frame containing, for each study, the effect size estimates
for each pathway from the SEM analyses. This data frame also contains meda-data
(e.g. sutdy species, study location, continent, life history traits of the species),
that are needed to fit the mixed-effect model.}

\item{Demog_rate}{Character specifying the level of the demographic rate on which
to subset the data. If NULL no subsetting by demographic rate is performed.}

\item{Trait_categ}{Character specifying the level of the trait on which to subset the data.}

\item{Clim}{Character specifying the level of the climatic variable on which to subset
the data.}

\item{Cov_fact}{A character specifying the name of the categorical variable to be included
as a fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case no
categorical variables are included and the overall global effect size is estimated.}

\item{COV}{A character specifying the name (or names, separated by '+') of the continuous
variable to be included as a fixed-effect covariate in the meta-analysis. Defaults to NULl,
in which case no continuous covariates are included and the overall global effect size
is estimated.}

\item{sel}{Character specifying the name to be included in the .pdf name, which indicates
the levels of demographic rate and trait for which the subsetting was done.}

\item{folder_name}{Character specifyng the path to the directory in which the results will be saved.}

\item{colr}{Vector specifying the colours to be used for the data points. The length of the
vector should correspond to the number of the levels in the categorical explanatory variable
included in the meta-analytical model, or should be one (if the single global effect size
across all studies is to be plotted).}

\item{DD}{Character specifying how to account for density dependence by
including population size as additional explanatory in the SEM. Possibilities
are: 'none' - no inclusion of density dependence, 'n_effectGR' - effect
of population size on growth rate, 'n_effectD' - effect of population size
on demographic rate, 'n_effectDGR' - effect of population size on demographic
rate and growth rate. Only options 'n_effectGR' and 'none' are possible for
simple structure of SEM, in which climate effect on GR i sonly mediated by trait
(i.e. \code{simpleSEM = TRUE}).}

\item{simpleSEM}{Logical (TRUE/FALSE) specifying that the simple SEM
structure will be fitted in which climate effect on GR is only mediated
by trait, demographic rates are not considered. Defaults to FALSE.}

\item{Trait}{Logical (TRUE/FALSE) specifying whether to include
the direct impact of trait on GR in the SEM fitted using a complex
structure, i.e. climate effect on GR is mediated by both trait and
demographic rate. Is applicable only if \code{simpleSEM = FALSE}.
Defaults to FALSE.}

\item{A}{A variance-covariance matrix based on phylogeny.}

\item{des.matrix}{A character specifying what type of design matrix to use if the factor
is included as one of the explanatory variables. By default treatment contrasts is used
 (des.matrix = "treatm.contrasts") but the user can also request identity matrix by
 specifying "identity".}

\item{all_relations}{Vector specifying the names of the relations from the SEMs that will
be used as response variables in meta-analyses to be fitted.}
}
\value{
Returns a tibble that includes one data frame and one tibble. A data frame
contains the estimated effect sizes, their standard errors, lower and upper confidence
intervals, their significance and the AIC for both the mixed-effects model fitted while
accounting and not for phylogenetic relatedness. This dataframe also has the column
'Relation' specifying for which relation the data were analyzed
(e.g. Demog_rate_mean<-Pop_mean', for more details see \code{\link{fit_meta_phylo}}).
A tibble contains three lists and one character variable. The first list contains estimated global
effect sizes, their standard errors, their significance and the AIC per each fitted mixed-effects model
defined by the 'Relation', as returned by \code{\link{fit_meta_phylo}}. The second list
contains data frames with the effect sizes and standard errors for each study.
The third list contains the heterogeneity metrics extracted from each meta-analysis.
Each data frame corresponds to the subset of the data per each 'Relation' type.
A character 'names' specifies the 'Relation' type.
}
\description{
\code{fit_all_meta} fits all mixed-effects meta-analytical models (per each relation
in the SEM), to extract global effect sizes across the studies in a specified subset
of data as defined by a demographic rate and a trait
}
\examples{
# prepare dataset, select only studies with phenological traits
dataPaths <- dataPaths \%>\%
                  dplyr::mutate(Species = dplyr::case_when(
                         Species == 'Cyanistes caeruleus' ~ 'Parus caeruleus',
                         Species == 'Thalasseus sandvicensis' ~ 'Sterna sandvicensis',
                         Species == 'Setophaga caerulescens' ~ 'Dendroica caerulescens',
                         Species == 'Thalassarche melanophris' ~ 'Thalassarche melanophrys',
                         Species == 'Ichthyaetus audouinii' ~ 'Larus audouinii',
                         Species == 'Stercorarius maccormicki' ~ 'Catharacta maccormicki',
                         TRUE ~ Species))

dataPaths$Species <- unlist(lapply(1:nrow(dataPaths), FUN = function(x){
  binary <- strsplit(as.character(dataPaths$Species[x]), " ")
  Underscore <- paste(binary[[1]][1], binary[[1]][2], sep = "_")}))
dataPaths$Sp_phylo <- dataPaths$Species

# fit the models for: Trait_mean<-det_Clim', 'Ind_GR<-det_Clim', 'Tot_GR<-det_Clim'
meta_Phen_Cov <- fit_all_meta(data_MA = dataPaths,
                              Demog_rate = NULL,
                              Trait_categ = 'Phenological',
                              Clim = 'Temperature',
                              Cov_fact = 'WeathQ',
                              COV = 'Pvalue',
                              sel = 'Temp_Phen_Cov',
                              folder_name = NULL,
                              colr = c('black', 'red'),
                              DD = 'n_effectGR',
                              simpleSEM = TRUE,
                              A = phyloMat,
                              all_Relations = c('Trait_mean<-det_Clim',
                              'Ind_GR<-det_Clim', 'Tot_GR<-det_Clim'))
meta_Phen_Cov

}
