% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_meta.R
\name{fit_meta}
\alias{fit_meta}
\title{Fit a meta-analytical model to assess the global effects across the studies}
\usage{
fit_meta(
  data_MA,
  Type_EfS = "Trait_mean<-det_Clim",
  Covar = NULL,
  COV = NULL,
  optimize = "nlminb",
  DD = "n_effectDGR"
)
}
\arguments{
\item{data_MA}{Data frame containing, for each study, the effect size estimates
for each pathway from the SEM analyses. This data frame also contains meda-data
(e.g. sutdy species, study location, continent, life history traits of the species),
that are needed to fit the mixed-effect model.}

\item{Type_EfS}{Character specifying which effect size type to use for a meta-analysis.
These types of effect sizes reflect different pathways in the fitted SEM. Possible are:
'Demog_rate_mean<-det_Clim', 'Demog_rate_mean<-Pop_mean', 'Demog_rate_mean<-Trait_mean',
'GR<-Demog_rate_mean',  'GR<-det_Clim', 'GR<-Pop_mean', 'Ind_DemRate<-det_Clim',
'Ind_GR<-det_Clim', 'Tot_DemRate<-det_Clim', 'Tot_GR<-det_Clim', and 'Trait_mean<-det_Clim'.}

\item{Covar}{Categorical specifying the name of the categorical variable to be included
as fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case no
categorical variables are included and the overall global effect size is estimated.}
}
\value{
Returns a tibble. If no categorical explanatory variables are included in the model,
then this tibble includes four data frames: 1. 'data' - the data frame with the estimate
of the global effect size, its standard error, its significance, the AIC of the fitted
mixed-effects model and the variances explained by random effects;
2. 'data_Efs' - the data frame with the effect sizes and their standard errors for each study;
3. 'data_R2' - the data frame that includes in addition to the effect size and standard error
per each study also R2 of the fitted relations; and 4. 'prop_data' the data frame with
the proportional contribution of the direct and indirect paths to the total path. Additionally,
the column 'names' contains the names of the fitted relation for each mixed-effects model.
If a categorical explanatory variable is included in the model, then the tibble
additionally contains the dara frame with the estimates of effect sizes (and their SEs)
for each level of the categorical variable; and the columns with the p value
indicating whether the categorical variable is significant, the AIC of the mixed-effects model
with categorical variable and the variances explained by random effects.
The AIC and the p values are obtained from the models fitted with ML, whereas
the parameter estimates from the models fitted using REML.
}
\description{
\code{fit_meta} fits mixed-effects meta-analytical model to extract a global
effect size across the studies
}
\examples{
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
check_noCovar <- fit_meta(data_MA = Coefs_Aut, Type_EfS = 'Trait_mean<-det_Clim',
                          Covar = NULL)
check_TraitCateg <- fit_meta(data_MA = Coefs_Aut, Type_EfS = 'Trait_mean<-det_Clim',
                             Covar = 'Trait_Categ')
check_TraitCateg
}
