% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mod.R
\name{fit_mod}
\alias{fit_mod}
\title{Fit structural equation model with requested specifications}
\usage{
fit_mod(
  biol_data,
  ID,
  DD = "none",
  weight = FALSE,
  correlation = FALSE,
  Trait = FALSE,
  simpleSEM = FALSE,
  ...
)
}
\arguments{
\item{biol_data}{Data frame with climate, trait and demographic rate
data for a given population and species.}

\item{ID}{Numeric giving a unique ID of the current dataset for
a given population and species.}

\item{DD}{Character specifying how to account for density dependence by
including population size as additional explanatory in the SEM. Possibilities
are: 'none' - no inclusion of density dependence, 'n_effectGR' - effect
of population size on growth rate, 'n_effectD' - effect of population size
on demographic rate, 'n_effectDGR' - effect of population size on demographic
rate and growth rate. Only options 'n_effectGR' and 'none' are possible for
simple structure of SEM, in which climate effect on GR i sonly mediated by trait
(i.e. \code{simpleSEM = TRUE}).}

\item{weight}{Logical (TRUE/FALSE) specifying whether to use inverse of
variances for traits and demographic rates (if \code{simpleSEM = FALSE})
as weights in the respective models. Defaults to FALSE.}

\item{correlation}{Logical (TRUE/FALSE) specifying whether to include
temporal autocorrelation with AR(1) structure. Defaults to FALSE.}

\item{Trait}{Logical (TRUE/FALSE) specifying whether to include
the direct impact of trait on GR in the SEM fitted using a complex
structure, i.e. climate effect on GR is mediated by both trait and
demographic rate. Is applicable only if \code{simpleSEM = FALSE}.
Defaults to FALSE.}

\item{simpleSEM}{Logical (TRUE/FALSE) specifying that the simple SEM
structure will be fitted in which climate effect on GR is only mediated
by trait, demographic rates are not considered. Defaults to FALSE.}

\item{...}{Additional parameters for the function \code{\link{fit_mod}}}
}
\value{
A fitted object of class psem (see
\code{\link[piecewiseSEM]{psem}} for more details
}
\description{
\code{fit_mod} fits structural equation model with
requested model structure for the specified study ID
}
\examples{

 # prepare the data to fit the model to one study only
 subs <- droplevels(dataSEM[dataSEM$ID == 7, ])
 full_NA <- data.frame(Year = seq(min(subs$Year), max(subs$Year), by = 1))
 consec_yrs <- merge(full_NA, subs, by = 'Year', all= TRUE)

 data_GR <- consec_yrs \%>\%
 dplyr::mutate(., Pop_mean_lag = c(Pop_mean[-1], NA)) \%>\%
 dplyr::mutate(., GR = log(Pop_mean_lag / Pop_mean)) \%>\%
 dplyr::filter(., !is.na(GR) & !is.na(Trait_mean) &
               !is.na(Demog_rate_mean) & !is.na(Pop_mean)) \%>\%
 dplyr::mutate(det_Clim = stats::resid(stats::lm(Clim ~ Year,
               data = .))) \%>\%
 dplyr::mutate(across(where(is.array), as.numeric))

# fit the model
test <- fit_mod(biol_data = data_GR, ID = 7,
                DD = 'n_effectGR', weight = TRUE,
                correlation = TRUE,
                Trait = TRUE,
                simpleSEM = TRUE)
}
