% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SEM.R
\name{fit_SEM}
\alias{fit_SEM}
\title{Fit structural equation model and return the results}
\usage{
fit_SEM(
  biol_data,
  ID,
  out_SEM,
  DD = "none",
  weight = FALSE,
  correlation = FALSE,
  standardize = FALSE,
  Trait = FALSE,
  simpleSEM = FALSE
)
}
\arguments{
\item{biol_data}{Data frame with climate, trait and demographic rate
data for a given population and species.}

\item{ID}{Numeric giving a unique ID of the current dataset for
a given population and species.}

\item{out_SEM}{Character specifying the library on the path where
the results of structural equation analysis are stored.}

\item{DD}{Character specifying how to account for density dependence by
including population size as additional explanatory in the SEM. Possibilities
are: 'none' - no inclusion of density dependence, 'n_effectGR' - effect
of population size on growth rate, 'n_effectD' - effect of population size
on demographic rate, 'n_effectDGR' - effect of population size on demographic
rate and growth rate. Only options 'n_effectGR' and 'none' are possible for
simple structure of SEM, in which climate effect on GR i sonly mediated by trait
(i.e. \code{simpleSEM = TRUE}).}

\item{weight}{Logical (TRUE/FALSE) specifying whether to use inverse of
variances for traits and demographic rates (if \code{simpleSEM = FALSE})
as weights in the respective models. Defaults to FALSE.}

\item{correlation}{Logical (TRUE/FALSE) specifying whether to include
temporal autocorrelation with AR(1) structure. Defaults to FALSE.}

\item{Trait}{Logical (TRUE/FALSE) specifying whether to include
the direct impact of trait on GR in the SEM fitted using a complex
structure, i.e. climate effect on GR is mediated by both trait and
demographic rate. Is applicable only if \code{simpleSEM = FALSE}.
Defaults to FALSE.}

\item{simpleSEM}{Logical (TRUE/FALSE) specifying that the simple SEM
structure will be fitted in which climate effect on GR is only mediated
by trait, demographic rates are not considered. Defaults to FALSE.}
}
\value{
A tibble with 33 columns that represent the metadata for a
given study ID and the results of SEM: Cstat, dTable, R2 of all
equations and path coefficients, as well as requested settings
for fitting SEM (whether weights, autocorrelation and traits are
included in the model).
}
\description{
\code{fit_SEM} fits structural equation model with
requested model structure for the specified study ID and returns
the results to be used in meta-analysis
}
\examples{
# fit SEM
mod_SEM <- fit_SEM(biol_data = dataSEM, ID = 1,
                out_SEM = 'output_SEM_all',
                DD = 'n_effectGR', weight = TRUE,
                correlation = TRUE,
                standardize = TRUE,
                Trait = FALSE,
                simpleSEM = TRUE)

}
