% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_paths.R
\name{ind_path}
\alias{ind_path}
\title{Provide estimates of median and SE of an indirect (focal) pathway by bootstrap}
\usage{
ind_path(
  x,
  y,
  z = NULL,
  x.se,
  y.se,
  z.se = NULL,
  numrep = 10000,
  omega = NULL,
  omega.se = NULL
)
}
\arguments{
\item{x}{Numeric reflecting the first path coefficient estimate
in the indirect path}

\item{y}{Numeric reflecting the second path coefficient estimate
in the indirect path}

\item{z}{Numeric reflecting the third path coefficient estimate
in the indirect path, defaults to NULL, in which case
the indirect path consists of two pathways.}

\item{x.se}{Numeric reflecting the standard error for
the first path coefficient estimate.}

\item{y.se}{Numeric reflecting the standard error for
the second path coefficient estimate.}

\item{z.se}{Numeric reflecting the standard error for
the third path coefficient estimate, defaults to NULL for
the case of the indirect path consisting of two pathways.}

\item{numrep}{Numeric for the number of replicates to use in a
bootstrap.}

\item{omega}{Numeric reflecting the fourth path coefficient estimate
(applicable only if Trait = TRUE in \code{fit_SEM}). Defaults to NULL.}

\item{omega.se}{Numeric reflecting the standard errir for the
fourth path coefficient estimate (applicable only if T
rait = TRUE in \code{fit_SEM}). Defaults to NULL.}
}
\value{
A dataframe that contains four columns: a median of the
bootstrap values as an estimate of an indirect path coefficient,
lower and upper confidence intervals, and SE of the estimate.
}
\description{
\code{ind_path} estimates median and SE of an indirect pathway
by bootstrap
}
\examples{
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
  forTrans <- subset(Coefs_Aut,
              select = c(Estimate,  Std.Error, Relation,
              Species, Location, ID)) \%>\%
              dplyr::rename(SError = Std.Error)
met_wide <- forTrans \%>\%
  tidyr::gather(variable, value, -(Relation:ID)) \%>\%
  tidyr::unite(temp, Relation, variable, sep = '/') \%>\%
  tidyr::spread(temp, value)

ind_p_Dem <- ind_path(x = met_wide$`Trait_mean<-det_Clim/Estimate`[1],
                  y = met_wide$`Demog_rate_mean<-Trait_mean/Estimate`[1],
                  z = NULL,
                  x.se = met_wide$`Trait_mean<-det_Clim/SError`[1],
                  y.se = met_wide$`Demog_rate_mean<-Trait_mean/SError`[1],
                  z.se = NULL, numrep = 10000)
ind_p_GR <- ind_path(x = met_wide$`Trait_mean<-det_Clim/Estimate`[2],
                     y = met_wide$`Demog_rate_mean<-Trait_mean/Estimate`[2],
                     z = met_wide$`GR<-Demog_rate_mean/Estimate`[2],
                     x.se = met_wide$`Trait_mean<-det_Clim/SError`[2],
                     y.se = met_wide$`Demog_rate_mean<-Trait_mean/SError`[2],
                     z.se = met_wide$`GR<-Demog_rate_mean/SError`[2],
                     numrep = 10000)

}
