% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest plot of effect sizes per each study and global effect sizes across studies}
\usage{
plot_forest(data_ES = check_DemCateg$data_EfS[[1]],
  data_globES = check_DemCateg$data_Covar[[1]], Covar = NULL,
  xlab = "Effect of temperature on trait", colr = c("blue", "green",
  "black"), pdf_basename = NULL, mar = c(4, 10, 2, 2),
  labels_ES = TRUE)
}
\arguments{
\item{data_ES}{Data frame containing, for each study, the effect size estimates and their
standard errors, returned as 'data_EfS' by the function \code{\link{fit_meta}}.}

\item{data_globES}{Data frame containing the global effect sizes across the studies,
returned by function {fit_meta} as either 'data_Covar' or 'data' (depending on whether
the explanatory categorical variable was included in the meta-analytical model).}

\item{Covar}{Categorical specifying the name of the categorical variable that was included
as fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case the
overall global effect size across all studies is plotted.}

\item{xlab}{Character specifying the label for the x axis.}

\item{colr}{Vector specifying the colours to be used for the data points. The length of the
vector should correspond to the number of the levels in the categorical explanatory variable
included in the meta-analytical model, or should be one (if the single global effect size
across all studies is to be plotted).}

\item{pdf_basename}{A character specifying the name of the .pdf for
the produced plot.}

\item{mar}{A vector specifying the plot margins, analogously to \code{\link[graphics]{par}}.}

\item{labels_ES}{Boolean specifying whether to plot the labels for all the single studies.
Defaults to TRUE.}
}
\value{
Plots a forest plot with effect sizes (and SEs) for each study,
and global effect size(s) on the bottom. The data used for the plot are
returned (invisibly).
}
\description{
\code{plot_forest} Plots a forest plot displaying the effect sizes
for each study and global effects across the studies
}
\examples{
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
check_noCovar <- fit_meta(data_MA = Coefs_Aut, Type_EfS = 'Trait_mean<-det_Clim',
                          Covar = NULL)
plot_forest(data_ES = check_noCovar$data_EfS[[1]],
            data_globES = check_noCovar$data[[1]],
            Covar = NULL, xlab = 'Effect of temperature on trait',
            colr = c('black'),
            pdf_basename = NULL,
            mar = c(4, 10, 2, 2),
            labels_ES = FALSE)
check_TraitCateg <- fit_meta(data_MA = Coefs_Aut, Type_EfS = 'Trait_mean<-det_Clim',
                             Covar = 'Trait_Categ')
plot_forest(data_ES = check_TraitCateg$data_EfS[[1]],
            data_globES = check_TraitCateg$data_Covar[[1]],
            Covar = 'Trait_Categ',
            xlab = 'Effect of temperature on trait',
            colr = c('red', 'black'),
            pdf_basename = NULL,
            mar = c(4, 10, 2, 2),
            labels_ES = TRUE)
}
