% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relation.R
\name{plot_relation}
\alias{plot_relation}
\title{Plot of the general relation (global effect size) and study-specific relations
(effect sizes)}
\usage{
plot_relation(
  fit_meta_data = meta_Phen_Repro_byTax,
  rel = 1,
  Covar = NULL,
  Demog_rate = "Survival",
  Trait_categ = "Phenological",
  Clim = "Temperature",
  sel = "Phen_surv",
  colr = c("black"),
  folder_name = "./output_overall/",
  mar = c(3, 4, 7, 2)
)
}
\arguments{
\item{fit_meta_data}{Tibble returned by the function \code{\link{fit_meta}}, that
contains in different data frames all the data necessary produce the plot.}

\item{rel}{Numeric specifying which relation (specified under the column 'names'
in the input tibble) to produce the plot for.}

\item{Covar}{Categorical specifying the name of the categorical variable that was included
as fixed-effect covariate in the meta-analysis. Defaults to NULL, in which case the
overall global effect size across all studies is plotted.}

\item{Demog_rate}{Character specifying the level of the demographic rate for which
analyses were conducted.}

\item{Trait_categ}{Character specifying the level of the trait, for which analyses were
conducted.}

\item{Clim}{Character specifying the level of the climatic variable, for which analyses
were conducted.}

\item{sel}{Character specifying the name to be included in the .pdf name, which indicates
the levels of demographic rate and trait for which the mixed-effect model was fitted.}

\item{colr}{Character vector specifying the colours to be used for the data points.
The length of the vector should correspond to the number of the levels in the
categorical explanatory variable included in the meta-analytical model, or should
be one (if the single global effect size across all studies is to be plotted).}

\item{folder_name}{Character specifyng the path to the directory in which
the results will be saved.}

\item{mar}{A vector specifying the plot margins, analogously to \code{\link[graphics]{par}}.}
}
\value{
Plots a relation between the dependent and independent variable
with a single line per effect size for each study, and global effect size(s)
together with their SE(s) shown with dashed line(s).
}
\description{
\code{plot_relation} Plots relations between independent and dependent variables
 displaying the effect sizes per each study and global effect across the studies
}
\examples{
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
meta_Phen_Repro <- fit_all_meta(data_MA = Coefs_Aut,
                                Clim = 'Temperature',
                                Demog_rate = 'Reproduction',
                                Trait_categ = 'Phenological',
                                tab = 'BirdType',
                                Covar = NULL,
                                sel = 'Phen_Repro',
                                folder_name = './output_overall/',
                                colr = c('black'))
plot_relation(fit_meta_data = meta_Phen_Repro,
                              rel = 1,
                              Covar = NULL,
                              Demog_rate = 'Reproduction',
                              Trait_categ = 'Phenological',
                              Clim = 'Temperature',
                              sel = 'Phen_Repro',
                              colr = c('black'),
                              folder_name  = './output_overall/',
                              mar = c(4, 4, 4, 2))

meta_Phen_Repro_byTax <- fit_all_meta(data_MA = Coefs_Aut,
                                     Demog_rate = 'Reproduction',
                                     Trait_categ = 'Phenological',
                                     Clim = 'Temperature',
                                     tab = 'Taxon',
                                     Covar = 'Taxon',
                                     sel = 'Phen_Repro',
                                     folder_name = './output_overall/',
                                     colr = c('black', 'green', 'red'))
plot_relation(fit_meta_data = meta_Phen_Repro_byTax,
                              rel = x,
                              Covar = 'Taxon',
                              Demog_rate = 'Reproduction',
                              Trait_categ = 'Phenological',
                              Clim = 'Temperature',
                              sel = 'Phen_Repro',
                              colr = c('black', 'green', 'red'),
                              folder_name  = './output_overall/',
                              mar = c(4, 4, 7, 2))

}
