% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uni_spSpec.R
\name{plot_uni_spSpec}
\alias{plot_uni_spSpec}
\title{Plot predictions of climate effects on traits, using species-specific predictors}
\usage{
plot_uni_spSpec(
  data_allEstim,
  mod_mv,
  lOut = 10,
  pdf_basename = NULL,
  xLab,
  yLab = "Estimate",
  byHemisphere = FALSE,
  miny = -1.2,
  maxy = 1.2
)
}
\arguments{
\item{data_allEstim}{A data frame with effect size estimates per relation for each
study (as returned by the SEM fitted with the function \code{\link{fit_SEM}}.}

\item{mod_mv}{Model object of class rma.mv fitted to test the
hypothesis about the effect of the specified species characteristic..}

\item{lOut}{Numeric specifying the length of the predictor variable, for which the
response of species traits to climate will be predicted and plotted.}

\item{pdf_basename}{A character specifying the name of the .pdf for
the produced plot.}

\item{xLab}{Character specifying the x axis label.}

\item{yLab}{Character specifying the y axis label.}

\item{byHemisphere}{Boolean specifying whether raw data points should be coloured depending on
which Hemisphere they are coming from.}

\item{miny}{Numeric specifying the minimum limit for the y axis.}

\item{maxy}{Numeric specifying the maximum limit for the y axis.}
}
\value{
Plots a forest plot with effect sizes (and SEs) for each study,
and global effect size(s) on the bottom. The data used for the plot are
returned (invisibly).
}
\description{
\code{plot_uni_spSpec} Plots predictions of the model using as a response effect
of climate on trait and as predictor one species-specific predictor at a time and
climate
}
\examples{
# prepare the dataset with phenological traits only
dataPaths_phen <- dataPaths \%>\%
    dplyr::filter(Trait_Categ == 'Phenological') \%>\%
    dplyr::mutate(absLat = abs(Latitude))
# fit the meta-analytical model
mod_CZ_PhenT_AbsLat <- metafor::rma.mv(Estimate ~ absLat + Pvalue,
                              V = Std.Error^2, random = list(~ 1|Species, ~1|ID, ~1|Location),
                              data = dataPaths_phen, method = 'ML')
plot_CZ_PhenT_AbsLat <- plot_uni_spSpec(data_allEstim = dataPaths_phen,
                                        mod_mv = mod_CZ_PhenT_AbsLat,
                                       lOut = 10, xLab = 'Absolute latitude',
                                       yLab = 'CZ estimate',
                                       pdf_basename = paste0(tempdir(), '/PlotCZ_PhenT_byAbsLat'),
                                       # attention: for this example we
                                       # write the data to a temporary directory,
                                       # to check its location type tempdir()
                                       byHemisphere = FALSE,
                                       miny = min(dataPaths_phen$Estimate) - 0.1,
                                       maxy = max(dataPaths_phen$Estimate) + 0.1)
message('Temporary directory is located at', tempdir())
message('Contents of the temporary directory after running fit_SEM()',
list.files(tempdir()))

}
