% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_paths.R
\name{tot_path}
\alias{tot_path}
\title{Provide estimates of median and SE of the total pathway by bootstrap}
\usage{
tot_path(
  direct,
  indir,
  ClDem = NULL,
  DemGR = NULL,
  direct.se,
  indir.se,
  ClDem.se = NULL,
  DemGR.se = NULL,
  numrep = 10000
)
}
\arguments{
\item{direct}{Numeric reflecting the estimate of the direct
path coefficient.}

\item{indir}{Numeric reflecting the estimate of the indirect
path coefficient.}

\item{ClDem}{Numeric reflecting the estimate of the path coefficient
between climate and demographic rate, in case the total path from
climate to growth rate is estimated. Defaults to NULL.}

\item{DemGR}{Numeric reflecting the estimate of the path coefficient
between demographic rate and growth rate,in case the total path from
climate to growth rate is estimated. Defaults to NULL.}

\item{direct.se}{Numeric reflecting the standard error for the direct
path coefficient.}

\item{indir.se}{Numeric reflecting the standard error for the indirect
path coefficient.}

\item{numrep}{Numeric for the number of replicates to use in a
bootstrap.}

\item{z.se}{Numeric reflecting the standard error for
the third path coefficient estimate, defaults to NULL for
the case of the indirect path consisting of two pathways.}
}
\value{
A dataframe that contains four columns: a median of the
bootstrap values as an estimate of an indirect path coefficient,
lower and upper confidence intervals, and SE of the estimate.
}
\description{
\code{tot_path} estimates median and SE of the total pathway
by bootstrap
}
\examples{
Coefs_Aut <- readRDS(file = './output_forSEM_temp/PathCoefs_allMods_Temp_Weights_DD_Autocor.RDS')
  forTrans <- subset(Coefs_Aut,
              select = c(Estimate,  Std.Error, Relation,
              Species, Location, ID)) \%>\%
              dplyr::rename(SError = Std.Error)
met_wide <- forTrans \%>\%
  tidyr::gather(variable, value, -(Relation:ID)) \%>\%
  tidyr::unite(temp, Relation, variable, sep = '/') \%>\%
  tidyr::spread(temp, value)

ind_p_Dem <- ind_path(x = met_wide$`Trait_mean<-det_Clim/Estimate`[1],
                  y = met_wide$`Demog_rate_mean<-Trait_mean/Estimate`[1],
                  z = NULL,
                  x.se = met_wide$`Trait_mean<-det_Clim/SError`[1],
                  y.se = met_wide$`Demog_rate_mean<-Trait_mean/SError`[1],
                  z.se = NULL, numrep = 10000)
tot_Dem <- tot_path(direct = met_wide$`Demog_rate_mean<-det_Clim/Estimate`[1],
                    indir = ind_p_Dem$Median,
                    ClDem = NULL,  DemGR= NULL,
                    direct.se = met_wide$`Demog_rate_mean<-det_Clim/SError`[1],
                    indir.se = ind_p_Dem$SE,
                    ClDem.se = NULL, DemGR.se = NULL,
                    numrep = 10000)

}
